+++ packages/expo-image-picker/src/ImagePicker.ts
@@ -1,4 +1,4 @@
-import { UnavailabilityError } from '@unimodules/core';
+import { UnavailabilityError, CodedError } from '@unimodules/core';
 import { PermissionStatus, PermissionExpiration } from 'unimodules-permissions-interface';
 
 import ExponentImagePicker from './ExponentImagePicker';
@@ -10,6 +10,37 @@ import {
   ImagePickerOptions,
   VideoExportPreset,
 } from './ImagePicker.types';
+
+function validateOptions(options: ImagePickerOptions) {
+  const { aspect, quality, videoMaxDuration } = options;
+
+  if (aspect != null) {
+    const [x, y] = aspect;
+
+    if (x <= 0 || y <= 0) {
+      throw new CodedError(
+        'ERR_INVALID_ARGUMENT',
+        `Invalid aspect ratio values ${x}:${y}. Provide positive numbers.`
+      );
+    }
+  }
+
+  if (quality && (quality < 0 || quality > 1)) {
+    throw new CodedError(
+      'ERR_INVALID_ARGUMENT',
+      `Invalid 'quality' value ${quality}. Provide a value between 0 and 1.`
+    );
+  }
+
+  if (videoMaxDuration && videoMaxDuration < 0) {
+    throw new CodedError(
+      'ERR_INVALID_ARGUMENT',
+      `Invalid 'videoMaxDuration' value ${videoMaxDuration}. Provide a non-negative number.`
+    );
+  }
+
+  return options;
+}
 
 export async function getCameraPermissionsAsync(): Promise<CameraPermissionResponse> {
   return ExponentImagePicker.getCameraPermissionsAsync();
@@ -33,7 +64,7 @@ export async function launchImageLibraryAsync(
   if (!ExponentImagePicker.launchImageLibraryAsync) {
     throw new UnavailabilityError('ImagePicker', 'launchImageLibraryAsync');
   }
-  return await ExponentImagePicker.launchImageLibraryAsync(options);
+  return await ExponentImagePicker.launchImageLibraryAsync(validateOptions(options));
 }
 
 export async function launchCameraAsync(
@@ -42,7 +73,7 @@ export async function launchCameraAsync(
   if (!ExponentImagePicker.launchCameraAsync) {
     throw new UnavailabilityError('ImagePicker', 'launchCameraAsync');
   }
-  return await ExponentImagePicker.launchCameraAsync(options);
+  return await ExponentImagePicker.launchCameraAsync(validateOptions(options));
 }
 
 export {